% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TR_aux_funs.R
\name{check_trailing_slash}
\alias{check_trailing_slash}
\title{Ensure trailing slash in directory path}
\usage{
check_trailing_slash(input_path)
}
\arguments{
\item{input_path}{A character string representing the file or directory path.}
}
\value{
The input path with a trailing slash, if it was not already present.
}
\description{
Checks if a given file path ends with a slash ("/"). If not, it appends one to
the path. This function is useful for preparing file paths for operations that
require directory paths to end with a slash, ensuring consistency and preventing
errors in file path manipulations.
}
