% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TR_aux_funs.R
\name{decimal_to_8bit}
\alias{decimal_to_8bit}
\title{Convert a decimal value to an 8-Bit Number}
\usage{
decimal_to_8bit(decimalValue)
}
\arguments{
\item{decimalValue}{A numeric value between 0 and 1 representing the decimal
value to be converted. Handling of values outside this
range is not specified.}
}
\value{
An integer between 0 and 255 representing the 8-bit
        version of the input decimal value.
}
\description{
Takes a decimal value (assumed to be between 0 and 1) and converts it
to an 8-bit number, ranging from 0 to 255. The input is scaled by 255
and rounded to produce the 8-bit representation.
}
