% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TR_proc_Flir_rgb_data.R
\name{flir_get_full_rgb}
\alias{flir_get_full_rgb}
\title{Extract a full uncropped RGB image from FLIR Radiometric JPEG}
\usage{
flir_get_full_rgb(
  img_path,
  out_path,
  replace = FALSE,
  exiftool_path,
  load_as = NULL
)
}
\arguments{
\item{img_path}{Path to the input FLIR radiometric JPEG file.}

\item{out_path}{Path where the extracted full RGB image will be saved.}

\item{replace}{Logical flag indicating whether an existing output file should
be replaced. Defaults to FALSE.}

\item{exiftool_path}{Path to the EXIFtool executable.}

\item{load_as}{(Optional) Specifies how to load the extracted image into R: "cimg"
for an `imager` package object, "SpatRaster" for a `terra` package
object, or NULL to only save the file without loading. Defaults to NULL.}
}
\value{
Returns TRUE if the extraction and optional loading were successful, FALSE
        otherwise. If `load_as` is specified, returns the loaded image object.
}
\description{
Extracts the full RGB image embedded within a FLIR radiometric JPEG file,
saving it to a specified output path. This function uses EXIFtool to extract
the embedded RGB image. Additionally, there's an option to load the extracted
image directly into R as either an `imager` package `cimg` object or a
`terra` package `SpatRaster` object, depending on the `load_as` parameter.
}
