% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TR_proc_Flir_thermal_data.R
\name{flir_raw_to_thermal}
\alias{flir_raw_to_thermal}
\title{Convert FLIR raw thermal image to a temperature SpatRaster object}
\usage{
flir_raw_to_thermal(
  img_path,
  exiftool_path,
  out_path = NULL,
  scale_factor = NULL,
  datatype = "FLT4S",
  overwrite = TRUE
)
}
\arguments{
\item{img_path}{Path to the input FLIR image file.}

\item{exiftool_path}{Path to the EXIFtool executable, used for extracting metadata
from the FLIR image.}

\item{out_path}{(Optional) Output file path (usually with .tif extension) where the
temperature raster file will be saved.}

\item{scale_factor}{(Optional) A numeric value to scale the temperature values.
Useful for converting temperature to a specific range or format
such as conversion to integer numbers for saving space.
If NULL, no scaling is applied.}

\item{datatype}{Data type for saving the raster file. Supported types include
"INT1S", "INT2S", "INT4S", "INT1U", "INT2U", "INT4U", "FLT4S",
and "FLT8S". Defaults to "FLT4S".}

\item{overwrite}{Logical flag indicating whether to overwrite the output raster file.
Defaults to TRUE.}
}
\value{
A `SpatRaster` object representing the temperature distribution in the
        input FLIR image. If `out_path` is TRUE, the raster is also saved
        to the specified output folder usually as a (Geo)TIFF file.
}
\description{
Converts raw thermal imagery from FLIR cameras into a temperature raster using
metadata extracted with EXIFtool. This function reads a FLIR JPEG image, extracts
necessary calibration parameters from its metadata, and computes the temperature
values for each pixel. The temperature data can be scaled by a factor and written
to a file as a `SpatRaster` object. This is useful for thermal image analysis,
allowing for the quantitative assessment of temperature distributions in the captured
scene.
}
