% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TR_proc_Flir_rgb_data.R
\name{flir_thumbnail_to_rast}
\alias{flir_thumbnail_to_rast}
\title{Convert FLIR camera thumbnail image to SpatRaster}
\usage{
flir_thumbnail_to_rast(img_path, as_decimal = FALSE)
}
\arguments{
\item{img_path}{Path to the FLIR image.}

\item{as_decimal}{A logical flag indicating whether to convert the pixel values
to decimal format. Defaults to FALSE.}
}
\value{
A `SpatRaster` object representing the FLIR thumbnail image.
}
\description{
Converts a thumbnail image extracted from a FLIR camera's radiometric JPEG
to a SpatRster object.
}
