% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TR_raster_filters.R
\name{laplacian_filter}
\alias{laplacian_filter}
\title{Apply a Laplacian filter for edge detection on a RGB SpatRaster object}
\usage{
laplacian_filter(rgb_rast, type = "simple")
}
\arguments{
\item{rgb_rast}{An RGB SpatRaster object on which the Laplacian filter will be applied.
The raster is expected to have three layers corresponding to the
RGB channels. This is converted to a grayscale image before
applying the Laplacian filter.}

\item{type}{A character string specifying the type of Laplacian kernel to use.
"simple" uses a kernel that focuses on vertical and horizontal
neighbors, while "diagonal" includes diagonal neighbors as well.
Defaults to "simple". Any other value will result in an error.}
}
\value{
A SpatRaster object representing the filtered image, highlighting the edges
        detected in the original RGB raster. This image emphasizes areas of rapid
        intensity change, corresponding to edges.
}
\description{
Applies the Laplacian filter to an RGB SpatRaster for edge detection, supporting
both simple and diagonal in the kernel. The function first converts
the RGB raster to a grayscale image by averaging the RGB values and scaling to
a [0, 1] range. Depending on the `type` parameter, a Laplacian kernel without
diagonal elements (option: 'simple') or with diagonal elements ('diagonal') is applied. This
highlights edges within the image by enhancing regions of rapid intensity change.
}
