% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TR_raster_aux_funs.R
\name{plot_temp_rast}
\alias{plot_temp_rast}
\title{Plot a temperature SpatRaster using a specified color palette}
\usage{
plot_temp_rast(rst, palette = "magma", scale_temp = NULL, ...)
}
\arguments{
\item{rst}{A raster object to be plotted.}

\item{palette}{A character string specifying the color palette to use.
Options include "magma", "plasma", "inferno" (from the viridis
package), "RdYlBu", "Spectral", "PuOr" (from the RColorBrewer
package), and "heat" for heat.colors. Defaults to "magma".}

\item{scale_temp}{A numeric value to scale the raster values by before plotting.
Defaults to 100.}

\item{...}{Additional arguments passed to the plot function.}
}
\description{
Plots a raster object using a specified color palette. This function allows for
the visualization of raster data with various color schemes to enhance the
representation of spatial patterns. The raster values are scaled by `scale_temp`
before plotting, and a range of palettes based on the viridis and RColorBrewer
packages are available. This function is useful for visualizing temperature
data or any other continuous variable represented in raster format.
}
