% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TR_thermal_super_res.R
\name{prepare_train_data}
\alias{prepare_train_data}
\title{Prepare training features from an RGB image}
\usage{
prepare_train_data(rgb_rst, temp_rst = NULL)
}
\arguments{
\item{rgb_rst}{A `SpatRaster` object containing the RGB SpatRaster data.}

\item{temp_rst}{A `SpatRaster` object containing the temperature SpatRaster data.
Default to NULL in case only prediction is to be performed. For
training this needs to be defined.}
}
\value{
A data frame where each row represents a pixel and columns represent the
        temperature (if temp is not NULL), original RGB values, spatial statistics for
        each RGB channel,and the results from Sobel and Laplacian edge detection.
        This data frame is suitable for use in machine learning model training/prediction.
}
\description{
Processes RGB (predictors) and temperature (target) raster data to generate a feature
dataset suitablefor training machine learning models. This function computes several spatial
statistics (mean and standard deviation) within multiple window sizes (3x3, 5x5,
7x7, 9x9) for each RGB channel. Additionally, it applies Sobel and Laplacian filters
for edge detection. The function ensures that the RGB and temperature rasters are
compatible in terms of spatial dimensions and resolution before proceeding.
}
