% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TR_aux_funs.R
\name{pretty_duration}
\alias{pretty_duration}
\title{Format duration (difftime) into a readable string}
\usage{
pretty_duration(duration)
}
\arguments{
\item{duration}{A `difftime` object representing the duration to be formatted.
The duration should be a positive value.}
}
\value{
A character string describing the duration in terms of minutes and
        seconds, formatted as "X minutes Y seconds".
}
\description{
Converts a duration object into a human-readable string format, showing
the duration in minutes and seconds. This function takes a `difftime`
object, converts it into seconds, and then formats it as "X minutes Y seconds".
This can be particularly useful for displaying time durations in a user-friendly manner.
}
