% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TR_raster_aux_funs.R
\name{rast_upsample}
\alias{rast_upsample}
\title{Upsample a SpatRaster to specified width and height dimensions}
\usage{
rast_upsample(rst, to_height = 1440, to_width = 1080, method = "near")
}
\arguments{
\item{rst}{A `SpatRaster` object representing the image to be upsampled.}

\item{to_height}{The target height (number of rows) for the upsampled image.
Defaults to 1440.}

\item{to_width}{The target width (number of columns) for the upsampled image.
Defaults to 1080.}

\item{method}{The method of interpolation to use for resampling. Supported
methods include "near" for nearest neighbor interpolation,
among others. Defaults to "near".}
}
\value{
A `SpatRaster` object representing the upsampled image at the specified
        target dimensions.
}
\description{
Upsamples a `SpatRaster` object to specified dimensions using a chosen
interpolation method. This function is designed to increase the spatial resolution
of an image by interpolating new pixel values in a manner consistent with the
specified resampling method. This is useful in contexts where image resolution
needs to be increased for analysis or visualization purposes.
}
