% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TR_raster_aux_funs.R
\name{raster_to_cimg}
\alias{raster_to_cimg}
\title{Convert SpatRaster to cimg Object}
\usage{
raster_to_cimg(x, interpolation = 0)
}
\arguments{
\item{x}{A `SpatRaster` object to be converted.}

\item{interpolation}{An integer indicating the interpolation method to use when
rotating the image.Zero (0) for nearest-neighbor interpolation, one (1) for linear.}
}
\value{
A `cimg` object corresponding to the input `SpatRaster`.
}
\description{
Transforms a `SpatRaster` object from the `terra` package into a `cimg`
object from the `imager` package. This conversion is useful for image processing
and analysis workflows that require the use of `imager` functions.
}
