% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TR_raster_aux_funs.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers from a SpatRaster based on percentile thresholds}
\usage{
remove_outliers(rast, pmin = 5, pmax = 100, na.rm = TRUE)
}
\arguments{
\item{rast}{A SpatRaster object from which outliers will be removed.}

\item{pmin}{The lower percentile used to define the lower threshold for outliers.
Values below this percentile will be set to NA. Defaults to 5.}

\item{pmax}{The upper percentile used to define the upper threshold for outliers.
Values above this percentile will be set to NA. Defaults to 95.}

\item{na.rm}{Remove NA's (default TRUE)}
}
\value{
A SpatRaster object similar to the input but with values outside the
        specified percentile thresholds set to NA.
}
\description{
This function removes outliers from a raster object by setting values outside
specified percentile thresholds to NA. It uses the quantile function to determine
the lower and upper bounds based on the `pmin` and `pmax` parameters, respectively.
Values below the lower percentile (`pmin`) or above the upper percentile (`pmax`)
are considered outliers and are set to NA. This can be useful for data
preprocessing or visualization, especially in spatial analysis and modeling,
to mitigate the impact of extreme values.
}
