% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TR_thermal_super_res.R
\name{rf_thermal_from_rgb}
\alias{rf_thermal_from_rgb}
\title{Predict thermal data from RGB image using Random Forest}
\usage{
rf_thermal_from_rgb(
  rgb_rst,
  temp_rst,
  rgb_high_res = NULL,
  npix = 10000,
  rm_cor = FALSE,
  rm_cor_thresh = 0.98,
  get_rf_model = TRUE,
  verbose = TRUE,
  scale_data = FALSE,
  ...
)
}
\arguments{
\item{rgb_rst}{An RGB `SpatRaster` object containing RGB data aligned with `temp_rst`.}

\item{temp_rst}{A `SpatRaster` object containing corresponding lower-resolution
thermal data.}

\item{rgb_high_res}{(Optional) A `SpatRaster` object containing high-resolution
RGB data for which thermal predictions are to be made.}

\item{npix}{The number of pixels to sample for model training.}

\item{rm_cor}{Logical flag indicating whether to remove highly correlated features
before model training.}

\item{rm_cor_thresh}{Correlation threshold for removing correlated features.}

\item{get_rf_model}{Logical flag indicating whether to return the Random Forest
model along with the predicted temperature raster(s).}

\item{verbose}{Logical flag for printing progress messages.}

\item{scale_data}{Logical flag indicating whether to scale the feature data before
model training.}

\item{...}{Additional arguments passed to the Random Forest model training function.}
}
\value{
Depending on `get_rf_model`, either a list containing the predicted
        temperature raster(s) and the Random Forest model, or just the predicted
        temperature raster(s). If `rgb_high_res` is provided, the list will include
        predictions for both the training resolution and the high-resolution RGB.
        List components include: `pred_temp_train` is the thermal image predicted
        for the training data (usually with lower resolution); `pred_temp_rgb_hr`
        is the predicted thermal data for the high-resolution RGB; `rf_mod` is the
        Random Forest model trained for the image.
}
\description{
Enhances the resolution or predicts thermal data based on high-resolution RGB images
and an existing low-resolution thermal raster using a Random Forest model. The model
is trained on features derived from the RGB and low-resolution thermal data, optionally
scaled and with high correlation features removed. The function can then predict
thermal data aligned with a provided high-resolution RGB image, effectively
super-resolving the thermal image or generating new thermal data where none exists.
}
