% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TR_proc_roboflow_masks.R
\name{simplify_roboflow_masks}
\alias{simplify_roboflow_masks}
\title{Simplify and aggregate ROI masks by label from Roboflow annotations}
\usage{
simplify_roboflow_masks(masks)
}
\arguments{
\item{masks}{A list generated by `get_roboflow_masks` containing the labels/categories (`labs`),
polygon masks (`pol_masks` as `SpatVector` objects), and raster masks (`rst_masks` as
`SpatRaster`) extracted from JSON annotations.}
}
\value{
A list similar in structure to the input but with all masks of the same label
        aggregated together. Each label will correspond to a single polygon mask and a
        single raster mask, with the latter indicating presence (1 or TRUE) or absence
        (`NA`) of the label at each pixel.
}
\description{
Aggregates polygon and raster masks by their labels from a list of masks,
produced by `get_roboflow_masks` function converting JSON polygon annotations to spatial formats.
This functioncombines polygons and raster masks of the same label into single, aggregated masks,
which is particularly useful for analyses that require handling each label as a single
entity or layer. The aggregation of raster masks involves creating a unified mask where
any pixel covered by at least one mask of a given label is marked as part of that label.
}
