% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TR_raster_filters.R
\name{sobel_filter}
\alias{sobel_filter}
\title{Apply Sobel filter for edge detection on RGB SpatRaster object}
\usage{
sobel_filter(rgb_rast)
}
\arguments{
\item{rgb_rast}{An RGB SpatRaster object on which the Sobel filter will be applied.
The raster is expected to have three layers corresponding to
the RGB channels.}
}
\value{
A SpatRaster object representing the magnitude of the edges detected in the
        original RGB raster. This raster has values ranging from 0 to a maximum
        gradient magnitude, indicating the strength of edges at each pixel.
}
\description{
This function applies the Sobel filter to an RGB SpatRaster for edge detection.
It first converts the RGB raster into a grayscale image by averaging the RGB
values and scaling to a [0, 1] range. Then, it applies Sobel operators (Gx and Gy)
for horizontal and vertical edge detection, respectively. The magnitude of the
gradient is computed, representing the edge strength at each pixel.
}
