% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TR_proc_Flir_rgb_data.R
\name{flir_rgb_to_rast}
\alias{flir_rgb_to_rast}
\title{Converts a FLIR image into an RGB SpatRaster}
\usage{
flir_rgb_to_rast(
  img_path,
  exiftool_path,
  out_path = NULL,
  crop = TRUE,
  crop_center_method = "scale_center_xy",
  out_path_crop = NULL,
  scale_factor = NULL,
  offset_x = NULL,
  offset_y = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{img_path}{Path to the input FLIR radiometric JPEG file.}

\item{exiftool_path}{Path to the EXIFtool executable, including trailing slash.}

\item{out_path}{Optional. The location/path where to save the full RGB image. If `NULL`
the directory of the input will be used with a file name suffix _full_RGB.jpg. This image
will be created by EXIFtool by extracting the RGB image from metadata. If a file
path is specified it must be of jpg extension.}

\item{crop}{Logical flag indicating whether to crop the extracted RGB image to
match the dimensions of the thermal image. Defaults to TRUE.}

\item{crop_center_method}{If crop=TRUE two methods are available for cropping the image:
\itemize{
   \item scale_center_xy" or 1 - this methods applies i) a scale factor (named `Real 2 IR` in
FLIR metadata) to increase the image, ii) determines the center of the re-scaled image by halving the image size,
iii) calculates the offset center by summing the X/Y offset parameters multiplied by the scale
parameter (i.e., `Real 2 IR`value), (iv) from the offset image center, extracts the low-resolution RGB overlapping
with the thermal image;

   \item scale_y_only or 2 - this methods applies i) a scale factor (named `Real 2 IR` in
FLIR metadata) to increase the image, ii) determines the center of the re-scaled image by halving the image size,
iii) calculates the offset center by summing the center x coordinate with the X offset value from metadata and the
center y coordinate with the Y Offset divided by the scale factor (`Real 2 IR`), (iv) from the offset image center,
extracts the low-resolution RGB overlapping with the thermal IR image;
}}

\item{out_path_crop}{Path to where the cropped imaged will be saved. This will use
`terra::writeRaster` to write the image data.}

\item{scale_factor}{The scale factor from FLIR metadata. Optionally it can be passed directly using this input
(default: not used equals NULL);}

\item{offset_x}{The scale X Offset from FLIR metadata. Optionally it can be passed directly using this input
(default: not used equals NULL);}

\item{offset_y}{The Y Offset from FLIR metadata. Optionally it can be passed directly using this input
(default: not used equals NULL);}

\item{overwrite}{Overwrite image data in `terra::writeRaster`? Default: TRUE}
}
\value{
A `SpatRaster` object representing the extracted image. If `crop=TRUE` the cropped
        RGB image will be returned instead of the full RGB image.
}
\description{
Extracts the full RGB image from a FLIR radiometric JPEG, optionally crops it
based on FLIR metadata to match the thermal image's dimensions, and converts
it to a `SpatRaster` object. This function can save the extracted and cropped
RGB image as a TIFF file. Cropping is useful to align the RGB image with the
corresponding thermal image, enabling direct comparison or fusion of data from
both sources.
}
